draft implementation: apply through safe apps sdk
