+++ packages/app/package.json
@@ -12,6 +12,7 @@
   },
   "dependencies": {
     "@headlessui/react": "^1.7.18",
+    "@safe-global/safe-apps-react-sdk": "^4.7.1",
     "@shazow/whatsabi": "^0.11.0",
     "@types/jest": "^29.5.5",
     "@types/node": "^20.9.0",
